requirements.txt
ipykernel
requests
snowflake-snowpark-python
boto3